+++ ibc/utils.go
@@ -45,7 +45,7 @@ func GetTransferSenderRecipient(packet channeltypes.Packet) (
 	return sender, recipient, data.Sender, data.Receiver, nil
 }
 
-// GetTransferAmount returns the amount from an ICS20 FungibleTokenPacketData.
+// GetTransferAmount returns the amount from an ICS20 FungibleTokenPacketData as a string.
 func GetTransferAmount(packet channeltypes.Packet) (string, error) {
 	// unmarshal packet data to obtain the sender and recipient
 	var data transfertypes.FungibleTokenPacketData
@@ -62,4 +62,61 @@ func GetTransferAmount(packet channeltypes.Packet) (string, error) {
 	}
 
 	return data.Amount, nil
+}
+
+// GetReceivedCoin returns the transferred coin from an ICS20 FungibleTokenPacketData
+// as seen from the destination chain.
+// If the receiving chain is the source chain of the tokens, it removes the prefix
+// path added by source (i.e sender) chain to the denom. Otherwise, it adds the
+// prefix path from the destination chain to the denom.
+func GetReceivedCoin(srcPort, srcChannel, dstPort, dstChannel, rawDenom, rawAmt string) sdk.Coin {
+	// NOTE: Denom and amount are already validated
+	amount, _ := sdk.NewIntFromString(rawAmt)
+
+	if transfertypes.ReceiverChainIsSource(srcPort, srcChannel, rawDenom) {
+		// remove prefix added by sender chain
+		voucherPrefix := transfertypes.GetDenomPrefix(srcPort, srcChannel)
+		unprefixedDenom := rawDenom[len(voucherPrefix):]
+
+		// coin denomination used in sending from the escrow address
+		denom := unprefixedDenom
+
+		// The denomination used to send the coins is either the native denom or the hash of the path
+		// if the denomination is not native.
+		denomTrace := transfertypes.ParseDenomTrace(unprefixedDenom)
+		if denomTrace.Path != "" {
+			denom = denomTrace.IBCDenom()
+		}
+
+		return sdk.Coin{
+			Denom:  denom,
+			Amount: amount,
+		}
+	}
+
+	// since SendPacket did not prefix the denomination, we must prefix denomination here
+	sourcePrefix := transfertypes.GetDenomPrefix(dstPort, dstChannel)
+	// NOTE: sourcePrefix contains the trailing "/"
+	prefixedDenom := sourcePrefix + rawDenom
+
+	// construct the denomination trace from the full raw denomination
+	denomTrace := transfertypes.ParseDenomTrace(prefixedDenom)
+	voucherDenom := denomTrace.IBCDenom()
+
+	return sdk.Coin{
+		Denom:  voucherDenom,
+		Amount: amount,
+	}
+}
+
+// GetSentCoin returns the sent coin from an ICS20 FungibleTokenPacketData.
+func GetSentCoin(rawDenom, rawAmt string) sdk.Coin {
+	// NOTE: Denom and amount are already validated
+	amount, _ := sdk.NewIntFromString(rawAmt)
+	trace := transfertypes.ParseDenomTrace(rawDenom)
+
+	return sdk.Coin{
+		Denom:  trace.IBCDenom(),
+		Amount: amount,
+	}
 }