+++ x/ibc/transfer/types/interfaces.go
@@ -1,0 +1,29 @@
+package types
+
+import (
+	"context"
+
+	"github.com/ethereum/go-ethereum/common"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
+	transfertypes "github.com/cosmos/ibc-go/v5/modules/apps/transfer/types"
+
+	erc20types "github.com/evmos/evmos/v10/x/erc20/types"
+)
+
+// BankKeeper defines the expected interface needed to check balances and send coins.
+type BankKeeper interface {
+	transfertypes.BankKeeper
+	HasBalance(ctx sdk.Context, addr sdk.AccAddress, amt sdk.Coin) bool
+}
+
+// ERC20Keeper defines the expected ERC20 keeper interface for supporting
+// ERC20 token transfers via IBC.
+type ERC20Keeper interface {
+	IsERC20Enabled(ctx sdk.Context) bool
+	IsERC20Registered(ctx sdk.Context, contractAddr common.Address) bool
+	GetTokenPairID(ctx sdk.Context, token string) []byte
+	GetTokenPair(ctx sdk.Context, id []byte) (erc20types.TokenPair, bool)
+	ConvertERC20(ctx context.Context, msg *erc20types.MsgConvertERC20) (*erc20types.MsgConvertERC20Response, error)
+}